requirements.txt
httpx>=0.27
SQLAlchemy>=2.0
pymysql>=1.1
python-dotenv>=1.0
rapidfuzz>=3.0
pandas>=2.2
streamlit>=1.39
PyYAML>=6.0.2
